refactored MemberDetail -> MemberActivity
