+++ docs/docs/auto-docs/components/MemberActivity/Card/EventsAttendedCardItem/functions/default.md
@@ -1,0 +1,19 @@
+[Admin Docs](/)
+
+***
+
+# Function: default()
+
+> **default**(`props`): `Element`
+
+Defined in: [src/components/MemberActivity/Card/EventsAttendedCardItem.tsx:26](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/components/MemberActivity/Card/EventsAttendedCardItem.tsx#L26)
+
+## Parameters
+
+### props
+
+[`InterfaceCardItem`](../interfaces/InterfaceCardItem.md)
+
+## Returns
+
+`Element`
