+++ docs/docs/auto-docs/components/MemberActivity/Card/EventsAttendedCardItem/interfaces/InterfaceCardItem.md
@@ -1,0 +1,86 @@
+[Admin Docs](/)
+
+***
+
+# Interface: InterfaceCardItem
+
+Defined in: [src/components/MemberActivity/Card/EventsAttendedCardItem.tsx:16](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/components/MemberActivity/Card/EventsAttendedCardItem.tsx#L16)
+
+Card component to display individual event attendance information
+Shows event details including title, date, location and organization
+
+## Param
+
+Organization ID
+
+## Param
+
+Event ID
+
+## Param
+
+Event start date
+
+## Param
+
+Event title
+
+## Param
+
+Event location
+
+## Properties
+
+### creator?
+
+> `optional` **creator**: `string`
+
+Defined in: [src/components/MemberActivity/Card/EventsAttendedCardItem.tsx:20](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/components/MemberActivity/Card/EventsAttendedCardItem.tsx#L20)
+
+***
+
+### eventId?
+
+> `optional` **eventId**: `string`
+
+Defined in: [src/components/MemberActivity/Card/EventsAttendedCardItem.tsx:22](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/components/MemberActivity/Card/EventsAttendedCardItem.tsx#L22)
+
+***
+
+### location?
+
+> `optional` **location**: `string`
+
+Defined in: [src/components/MemberActivity/Card/EventsAttendedCardItem.tsx:21](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/components/MemberActivity/Card/EventsAttendedCardItem.tsx#L21)
+
+***
+
+### orgId?
+
+> `optional` **orgId**: `string`
+
+Defined in: [src/components/MemberActivity/Card/EventsAttendedCardItem.tsx:23](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/components/MemberActivity/Card/EventsAttendedCardItem.tsx#L23)
+
+***
+
+### startdate?
+
+> `optional` **startdate**: `string`
+
+Defined in: [src/components/MemberActivity/Card/EventsAttendedCardItem.tsx:19](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/components/MemberActivity/Card/EventsAttendedCardItem.tsx#L19)
+
+***
+
+### time?
+
+> `optional` **time**: `string`
+
+Defined in: [src/components/MemberActivity/Card/EventsAttendedCardItem.tsx:18](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/components/MemberActivity/Card/EventsAttendedCardItem.tsx#L18)
+
+***
+
+### title
+
+> **title**: `string`
+
+Defined in: [src/components/MemberActivity/Card/EventsAttendedCardItem.tsx:17](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/components/MemberActivity/Card/EventsAttendedCardItem.tsx#L17)
