--- docs/docs/auto-docs/components/MemberActivity/EventsAttendedCardItem/functions/default.md
@@ -1,19 +1,0 @@
-[Admin Docs](/)
-
-***
-
-# Function: default()
-
-> **default**(`props`): `Element`
-
-Defined in: [src/components/MemberActivity/EventsAttendedCardItem.tsx:26](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/components/MemberActivity/EventsAttendedCardItem.tsx#L26)
-
-## Parameters
-
-### props
-
-[`InterfaceCardItem`](../interfaces/InterfaceCardItem.md)
-
-## Returns
-
-`Element`
