+++ docs/docs/auto-docs/components/MemberActivity/MemberActivityMocks/variables/errorMocks.md
@@ -1,0 +1,31 @@
+[Admin Docs](/)
+
+***
+
+# Variable: errorMocks
+
+> `const` **errorMocks**: `object`[]
+
+Defined in: [src/components/MemberActivity/MemberActivityMocks.ts:42](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/components/MemberActivity/MemberActivityMocks.ts#L42)
+
+## Type declaration
+
+### error
+
+> **error**: `Error`
+
+### request
+
+> **request**: `object`
+
+#### request.query
+
+> **query**: `DocumentNode` = `EVENT_DETAILS`
+
+#### request.variables
+
+> **variables**: `object`
+
+#### request.variables.id
+
+> **id**: `string` = `'event123'`
