+++ src/components/MemberActivity/Card/EventsAttendedCardItem.spec.tsx
