+++ src/components/MemberActivity/Modal/CustomCell/customTableCell.tsx
@@ -2,7 +2,7 @@ import { useQuery } from '@apollo/client';
 import { CircularProgress, TableCell, TableRow } from '@mui/material';
 import { EVENT_DETAILS } from 'GraphQl/Queries/Queries';
 import React from 'react';
-import styles from '../../style/app.module.css';
+import styles from '../../../../style/app.module.css';
 import { Link } from 'react-router-dom';
 /**
  * Custom table cell component to display event details
